impl'd functions to execute wgetpaste & manage the process
These functions create an async wgetpaste process, and depending on
whether you pass an ":stdin" or ":files" argument, sends the input
over a pipe or just passes the files to operate on via arguments.

We need to make sure to send a newline after writing the ":stdin"
string data if there is none in the data otherwise the process seems
to hang for some reason!

The sentinel function is what ends up processing the callback
functions after the process exits. It calls
"wgetpaste-process-success" if the process returns a zero exit status,
and this function could be overriden by users.

Maybe we will add a hook to be called when the process exits as well
later on?